requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
cssselect==1.1.0
fake-useragent==0.1.11
flaky==3.7.0
idna==3.4
importlib-metadata==5.0.0
lxml==4.9.1
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
pyTelegramBotAPI==4.7.0
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
w3lib==2.0.1
websockets==10.3
zipp==3.9.0
